got it working with strings

+++ tests/u01/cstub.h
@@ -1,7 +1,7 @@
 #ifndef CSTUB_H_
 #define CSTUB_H_
 
-unsigned int rng_2x32 (void);
+unsigned long rng_2x32 (void);
 void initialize_randii (char**);
 
 #endif // CSTUB_H_
