+++ tests/u01/main.c
@@ -3,7 +3,7 @@
 
 int main(int argc, char** argv) {
   initialize_randii(argv);
-  const unsigned int i = rng_2x32();
-  printf("\n%d\n", i);
+  const unsigned long i = rng_2x32();
+  printf("\n%lu\n", i);
   return 0;
 }
